refactor(backend): extract agent-service helpers into domain modules
