+++ src/backend/agent/account-usage.ts
@@ -1,0 +1,39 @@
+import type { AccountUsage, AccountUsageLimit, AccountUsageWindow } from "@openbot/contracts/ipc";
+import { isNumber } from "@openbot/contracts/runtime-values";
+import type { AccountRateLimitResult, AccountRateLimitsReadResult } from "../protocol";
+
+export function normalizeAccountUsage(rateLimits: AccountRateLimitsReadResult | null): AccountUsage {
+  const entries = rateLimits?.rateLimitsByLimitId
+    ? Object.entries(rateLimits.rateLimitsByLimitId).filter((entry): entry is [string, AccountRateLimitResult] =>
+        Boolean(entry[1]),
+      )
+    : [];
+  if (entries.length === 0 && rateLimits?.rateLimits) {
+    entries.push([rateLimits.rateLimits.limitId ?? "codex", rateLimits.rateLimits]);
+  }
+  const limits = entries.map(([id, limit]) => normalizeAccountLimit(id, limit));
+
+  return { limits };
+}
+
+export function normalizeAccountLimit(id: string, limit: AccountRateLimitResult): AccountUsageLimit {
+  return {
+    id: limit.limitId ?? id,
+    primary: normalizeUsageWindow(limit.primary),
+    secondary: normalizeUsageWindow(limit.secondary),
+  };
+}
+
+export function normalizeUsageWindow(window: AccountRateLimitResult["primary"]): AccountUsageWindow | null {
+  const usedPercent = finiteNumberOrNull(window?.usedPercent);
+  if (usedPercent === null) return null;
+  return {
+    usedPercent: Math.max(0, Math.min(100, usedPercent)),
+    windowDurationMins: finiteNumberOrNull(window?.windowDurationMins),
+    resetsAt: finiteNumberOrNull(window?.resetsAt),
+  };
+}
+
+export function finiteNumberOrNull(value: unknown): number | null {
+  return isNumber(value) && Number.isFinite(value) ? value : null;
+}
